added module to check for open file shares

+++ lib/modules/default-ssh.py
@@ -48,6 +48,8 @@ class Module(BaseModule):
 
     def run(self, inventory):
 
+        self.check_progs()
+
         # write targets to file
         with open(self.targets_file, 'w') as f:
             for host in inventory:
@@ -55,52 +57,57 @@ class Module(BaseModule):
                     self.num_targets += 1
                     f.write(host['IP Address'] + '\n')
 
-        try:
+        if self.num_targets <= 0:
+            print('\n[+] No valid targets for Patator scan')
 
-            patator_command = ['patator', 'ssh_login', '--threads={}'.format(self.threads), \
-                'user=COMBO00', 'password=COMBO01', 'host=FILE1', 
-                '--max-retries=2', '0={}'.format(self.creds_file),\
-                '1={}'.format(self.targets_file)]
+        else:
 
-            print('\n[+] Running patator against {:,} targets:\n\t> {}\n'.format(self.num_targets, ' '.join(patator_command)))
-
-            if self.patator_process is None:
-                self.patator_process = sp.Popen(patator_command, stdout=sp.PIPE, stderr=sp.PIPE)
-                sleep(2)
-
-                with open(self.patator_valid_creds, 'w') as valid_creds_file:
-                    with open(self.patator_log_file, 'w') as log_file:
-                        for line in io.TextIOWrapper(self.patator_process.stderr, encoding='utf-8'):
-                            # pass through stdout to log
-                            log_file.write(line)
-                            line = ''.join(line.split('patator')[1:]).strip()
-                            print('\r{}'.format(line), end='')
-                            if 'INFO - 0' in line:
-                                valid_creds_file.write(line)
-                                print('\r' + line)
-                                #try:
-                                cred_str = line.split()[6]
-                                creds = ':'.join(cred_str.split(':')[:2])
-                                ip = ipaddress.ip_address(cred_str.split(':')[-1])
-                                inventory.hosts[ip].update({'Default SSH Login': creds})
-                                #except ValueError:
-                                #    continue
-
-                self.patator_process = None
-
-            else:
-                raise PatatorError('Patator is already running')
-
-        except KeyboardInterrupt:
-            print('\n\n[!] Cancelling Patator scan, please wait')
             try:
-                self.patator_process.send_signal(SIGINT)
-                sleep(1)
-                self.patator_process.send_signal(SIGINT)
-                sleep(1)
-                self.patator_process.terminate()
-            except AttributeError:
-                pass
+
+                patator_command = ['patator', 'ssh_login', '--threads={}'.format(self.threads), \
+                    'user=COMBO00', 'password=COMBO01', 'host=FILE1', 
+                    '--max-retries=2', '0={}'.format(self.creds_file),\
+                    '1={}'.format(self.targets_file)]
+
+                print('\n[+] Running patator against {:,} targets:\n\t> {}\n'.format(self.num_targets, ' '.join(patator_command)))
+
+                if self.patator_process is None:
+                    self.patator_process = sp.Popen(patator_command, stdout=sp.PIPE, stderr=sp.PIPE)
+                    sleep(2)
+
+                    with open(self.patator_valid_creds, 'w') as valid_creds_file:
+                        with open(self.patator_log_file, 'w') as log_file:
+                            for line in io.TextIOWrapper(self.patator_process.stderr, encoding='utf-8'):
+                                # pass through stdout to log
+                                log_file.write(line)
+                                line = ''.join(line.split('patator')[1:]).strip()
+                                print('\r{}'.format(line), end='')
+                                if 'INFO - 0' in line:
+                                    valid_creds_file.write(line)
+                                    print('\r' + line)
+                                    #try:
+                                    cred_str = line.split()[6]
+                                    creds = ':'.join(cred_str.split(':')[:2])
+                                    ip = ipaddress.ip_address(cred_str.split(':')[-1])
+                                    inventory.hosts[ip].update({'Default SSH Login': creds})
+                                    #except ValueError:
+                                    #    continue
+
+                    self.patator_process = None
+
+                else:
+                    raise PatatorError('Patator is already running')
+
+            except KeyboardInterrupt:
+                print('\n\n[!] Cancelling Patator scan, please wait')
+                try:
+                    self.patator_process.send_signal(SIGINT)
+                    sleep(1)
+                    self.patator_process.send_signal(SIGINT)
+                    sleep(1)
+                    self.patator_process.terminate()
+                except AttributeError:
+                    pass
